Revert "Exam Percentage Change Seconds to Minutes"
This reverts commit 3671b79578053b2a8820eff9e8dc219c8a918f83.

+++ AndersonExamData/IDExaminee.cs
@@ -9,7 +9,7 @@ namespace AndersonExamData
         #endregion
 
         #region READ
-       int Percentage(int examineeId);
+        decimal Percentage(int examineeId);
         #endregion
 
         #region UPDATE
